add: double speed on firing strategies, movement strategies on player movement

+++ src/main/java/com/shootemup/g53/controller/firing/MovingBulletStrategy.java
@@ -11,7 +11,7 @@ public class MovingBulletStrategy extends FiringStrategy {
     private MovementStrategy movementStrategy;
 
 
-    public MovingBulletStrategy(MovementStrategy movementStrategy, int speed, int fireRate) {
+    public MovingBulletStrategy(MovementStrategy movementStrategy, double speed, int fireRate) {
         super();
         this.movementStrategy = movementStrategy;
         this.bulletSpeed = speed;
