+++ src/main/java/com/shootemup/g53/model/element/Bullet.java
@@ -1,6 +1,5 @@
 package com.shootemup.g53.model.element;
 
-import com.shootemup.g53.controller.movement.MovementStrategy;
 import com.shootemup.g53.model.util.Position;
 import com.shootemup.g53.model.util.objectpool.PoolableObject;
 
@@ -9,7 +8,7 @@ public class Bullet extends MovableElement implements PoolableObject {
     private int size;
     private boolean active = false;
 
-    public Bullet(Position position, String color,int speed, int size) {
+    public Bullet(Position position, String color, double speed, int size) {
         super(position, color, speed);
         this.size = size;
         this.active = true;
@@ -24,7 +23,7 @@ public class Bullet extends MovableElement implements PoolableObject {
         return new Bullet(new Position(position.getX(), position.getY()), getColor(),getSpeed(), size);
     }
 
-    public void init(int x, int y, String color, int size,int speed) {
+    public void init(int x, int y, String color, int size,double speed) {
         position.setX(x);
         position.setY(y);
         setColor(color);
